Move WeeklyXPRequestForm from Accounts to Game Fixes #776
